Add length to the OAuth2 state (#152)
The OAuth2 provider we are working with returns an error before showing the username/password login screen due to the state variable not having enough length / entropy.  This happens if you navigate to the root URL "/" or any URL path that is less than 8 characters long.  Navigating to "/clusters/local" or another longer path allows the login to function as normal.

This PR makes the state query param a JSON object with the redirect URL in it, simply since I couldn't think of a better way to make it longer 🤷‍♀️.  The length of the JSON key is long enough so that our OAuth2 provider will properly redirect even on the shortest URL "/".  I'm happy to use a different scheme if you have any ideas.

I've also added some basic unit tests around the OAuth2 logic to make sure this functionality all works.

Co-authored-by: Andrew Russell <russell@idelic.com>
Reviewed-on: https://codeberg.org/hjacobs/kube-web-view/pulls/152
Co-authored-by: idelic_russell <idelic_russell@noreply.codeberg.org>
Co-committed-by: idelic_russell <idelic_russell@noreply.codeberg.org>