Add parameter to disable access_log (#151)
I'm working on a patch to allow disabling access logs. This is because they expose JWT tokens which get logged and it raises alerts for us. This is a very rudimentary approach I've added here to be able to disable only the access_logs while keeping the rest of the container logs which might be useful for users.

Please let me know what you think about it and how this can be improved. I'll then also add updates for the documentation.

Co-authored-by: Noor Malik <nmalik@zalando-70804.corp.ad.zalando.net>
Reviewed-on: https://codeberg.org/hjacobs/kube-web-view/pulls/151
Co-authored-by: demoncoder95 <demoncoder95@noreply.codeberg.org>
Co-committed-by: demoncoder95 <demoncoder95@noreply.codeberg.org>